+++ server/src-lib/Hasura/Backends/MSSQL/Connection.hs
@@ -4,6 +4,7 @@ import           Hasura.Prelude
 
 import qualified Data.Environment        as Env
 import qualified Data.Pool               as Pool
+import qualified Data.Text               as T
 import qualified Database.ODBC.SQLServer as ODBC
 
 import           Control.Exception
@@ -14,7 +15,6 @@ import           Data.Aeson.TH
 import           Data.Text               (pack, unpack)
 import           Hasura.Base.Error
 import           Hasura.Incremental      (Cacheable (..))
-
 
 -- | ODBC connection string for MSSQL server
 newtype MSSQLConnectionString
@@ -167,3 +167,6 @@ instance Cacheable MSSQLSourceConfig where
 
 instance ToJSON MSSQLSourceConfig where
   toJSON = toJSON . _mscConnectionString
+
+newtype MSSQLConnErr = MSSQLConnErr { getConnErr :: T.Text }
+  deriving (Show, Eq, ToJSON)